+++ vector/src/main/java/im/vector/activity/VectorRoomActivity.java
@@ -3212,9 +3212,15 @@ public class VectorRoomActivity extends MXCActionBarActivity implements
                     VectorUtils.getRoomDisplayName(this, mSession, mRoom), mRoom.getState().getMemberName(member.mSender)));
         }
 
+        // On mobile side, the modal to allow to add a reason to ban/kick someone isn't yet implemented
+        // That's why, we don't display the TextView "Motif :" for now.
         TextView subInvitationTextView = findViewById(R.id.room_preview_subinvitation_textview);
-        subInvitationTextView.setText(getString(R.string.reason_colon, member.reason));
-
+        if (!TextUtils.isEmpty(member.reason)) {
+            final String reason = getString(R.string.reason_colon, member.reason);
+            subInvitationTextView.setText(reason);
+        } else {
+            subInvitationTextView.setText(null);
+        }
 
         Button joinButton = findViewById(R.id.button_join_room);